upload map files with hash
