+++ package.json
@@ -14,7 +14,7 @@
     "storybook": "start-storybook -p 9001",
     "watch": "webpack-dev-server --config ./packages/server/webpack.config.dev.js --colors --display-error-details",
     "build": "webpack --config ./packages/server/webpack.config.prod.js --output-path=`pwd`",
-    "bugsnag:release": "node ./scripts/bugsnag-release.js"
+    "bugsnag:release": "node ./packages/server/scripts/bugsnag-release.js"
   },
   "repository": {
     "type": "git",
