+++ packages/server/package.json
@@ -18,7 +18,6 @@
     "@bufferapp/shutdown-helper": "0.2.0",
     "@bugsnag/js": "^5.0.1",
     "body-parser": "1.17.2",
-    "bugsnag": "1.11.0",
     "cookie-parser": "1.4.3",
     "express": "4.15.3",
     "jsonwebtoken": "7.4.1",
