+++ packages/server/scripts/bugsnag-release.js
@@ -1,16 +1,6 @@
 const reportBuild = require('bugsnag-build-reporter');
-const fs = require('fs');
 
-const bugsnagApiKey = '6d235b284e8baf3c8d669a1991844969';
-
-const getAppReleaseData = () => {
-  try {
-    const versionJson = JSON.parse(fs.readFileSync('/src/version.json', 'utf8'));
-    return versionJson;
-  } catch (e) {
-    return null;
-  }
-};
+const { bugsnagApiKey, getAppReleaseData } = require('./bugsnag-config');
 
 const releaseData = getAppReleaseData();
