+++ packages/server/webpack.config.prod.js
@@ -1,14 +1,18 @@
+const webpack = require('webpack');
 const merge = require('webpack-merge');
 // const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 const common = require('./webpack.config.common.js');
 
 const merged = merge(common, {
   mode: 'production',
-  devtool: 'source-map',
+  devtool: undefined,
   optimization: {
     nodeEnv: 'production',
   },
-  // plugins: [new BundleAnalyzerPlugin()],
+  plugins: [new webpack.SourceMapDevToolPlugin({
+    filename: '[name].[hash].js.map',
+    exclude: ['bundle.css'],
+  })],
 });
 
 module.exports = merged;
